Add snapping hotkey to edit toolbar
Added a hotkey 's' to snap in the EditToolbar. Note - included a new library (react-hotkeys-hook) to do this. Changed the Toolbar component to a functional component so we could use react-hotkeys-hook and not the older version which no longer recieves updates. react-hotkeys-hook takes care of scoping the hotkey and dealing with not applying it when focuses on a textbox (e.g. search) and similar